requirements.txt
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
requests==2.14.2
